+++ lib/src/ports.dart
@@ -2,24 +2,17 @@ import "dart:isolate";
 import "dart:async";
 
 /// A type-safe [SendPort] that only allows you to send [T] objects.
-class TypedSendPort<T> {
-  final SendPort _port;
-
-  /// Wraps a native [SendPort].
-  TypedSendPort(this._port);
-
+extension type TypedSendPort<T>(SendPort _port) {
   /// Sends a [T] object.
   void send(T obj) => _port.send(obj);
 }
 
 /// A type-safe [ReceivePort] that only receives [T] objects.
-class TypedReceivePort<T> extends StreamView<T> {
-  final ReceivePort _port;
+extension type TypedReceivePort<T>(ReceivePort _port) {
+  /// A broadcast stream with all incoming data.
+  Stream<T> get stream => _port.asBroadcastStream().cast();
 
-  /// Wraps a native [ReceivePort] and casts it to [T].
-  TypedReceivePort(this._port) : super(_port.cast<T>());
-
-  /// Get the [TypedSendPort] for this [ReceivePort].
+  /// Get the type-safe [SendPort] for this [ReceivePort].
   TypedSendPort<T> get sendPort => TypedSendPort<T>(_port.sendPort);
 
   /// Closes the underlying port by calling [ReceivePort.close].